Refactor Date & Time Classes
