+++ src/test/java/seedu/address/logic/commands/lesson/AddToLessonCommandTest.java
@@ -15,6 +15,8 @@ import javafx.collections.ObservableList;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.datetime.Date;
+import seedu.address.model.datetime.Time;
 import seedu.address.model.lesson.Lesson;
 import seedu.address.model.student.Name;
 import seedu.address.model.student.Student;
@@ -72,8 +74,8 @@ public class AddToLessonCommandTest {
     private class ModelStubWithLesson extends ModelStub {
 
         private final Lesson lesson = new Lesson(
-                new seedu.address.model.consultation.Date("2024-10-20"),
-                new seedu.address.model.consultation.Time("14:00"),
+                new Date("2024-10-20"),
+                new Time("14:00"),
                 FXCollections.observableArrayList(),
                 Map.of());
