+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -40,8 +40,8 @@ import seedu.address.logic.commands.lesson.ListLessonsCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.consultation.Consultation;
-import seedu.address.model.consultation.Date;
-import seedu.address.model.consultation.Time;
+import seedu.address.model.datetime.Date;
+import seedu.address.model.datetime.Time;
 import seedu.address.model.lesson.Lesson;
 import seedu.address.model.student.IsStudentOfCoursePredicate;
 import seedu.address.model.student.Name;
