+++ src/test/java/seedu/address/logic/parser/lesson/AddLessonCommandParserTest.java
@@ -10,8 +10,8 @@ import java.util.Map;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.lesson.AddLessonCommand;
-import seedu.address.model.consultation.Date;
-import seedu.address.model.consultation.Time;
+import seedu.address.model.datetime.Date;
+import seedu.address.model.datetime.Time;
 import seedu.address.model.lesson.Lesson;
 
 public class AddLessonCommandParserTest {
