+++ src/test/java/seedu/address/model/datetime/DateTest.java
@@ -1,9 +1,12 @@
-package seedu.address.model.consultation;
+package seedu.address.model.datetime;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 
 import org.junit.jupiter.api.Test;
 
@@ -25,6 +28,9 @@ public class DateTest {
         String validDate = "2024-10-19";
         Date date = new Date(validDate);
         assertEquals(validDate, date.getValue());
+
+        LocalDate localDate = LocalDate.parse(validDate, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+        assertEquals(localDate, date.getLocalDateValue());
     }
 
     @Test
