+++ src/test/java/seedu/address/storage/JsonAdaptedConsultationTest.java
@@ -12,8 +12,8 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.consultation.Consultation;
-import seedu.address.model.consultation.Date;
-import seedu.address.model.consultation.Time;
+import seedu.address.model.datetime.Date;
+import seedu.address.model.datetime.Time;
 import seedu.address.model.student.Student;
 import seedu.address.testutil.ConsultationBuilder;
 import seedu.address.testutil.StudentBuilder;
